Title: 산타의 선물 공장, Time: 53ms, Memory: 8MB, Status: Runtime Error - CodeTree

+++ 240411/산타의 선물 공장/santa-gift-factory.cpp
@@ -11,7 +11,7 @@ struct Node {
 };
 
 Node nodes[100001];
-unordered_map<int, int> id_map;
+unordered_map<long long, long long> id_map;
 
 struct List {
 	Node head;
